+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/model/DocUpdateSentinel.java
@@ -189,7 +189,7 @@ public class DocUpdateSentinel
                   {
                      // We're quitting. Save one last time.
                      final Token token = event.acquire();
-                     boolean saving = doSave(null, null, null,
+                     boolean saving = doSave(null, null, null, true,
                            new ProgressIndicator()
                      {
                         public void onProgress(String message)
@@ -233,15 +233,21 @@ public class DocUpdateSentinel
 
    public void withSavedDoc(final Command onSaved)
    {
-      withSavedDoc(onSaved, null);
+      withSavedDocImpl(true, onSaved, null);
    }
 
-   public void withSavedDoc(final Command onSaved,
+   public void withSavedDocNoRetry(final Command onSaved)
+   {
+      withSavedDocImpl(false, onSaved, null);
+   }
+
+   private void withSavedDocImpl(final boolean retryWrite,
+                                 final Command onSaved,
          final CommandWithArg<String> onError)
    {
       if (changeTracker_.hasChanged())
       {
-         boolean saved = doSave(null, null, null, new ProgressIndicator() {
+         boolean saved = doSave(null, null, null, retryWrite, new ProgressIndicator() {
 
             @Override
             public void onCompleted()
@@ -273,7 +279,7 @@ public class DocUpdateSentinel
    {
       if (changeTracker_.hasChanged())
       {
-         return doSave(null, null, null, new ProgressIndicator()
+         return doSave(null, null, null, false, new ProgressIndicator()
          {
 
             @Override
@@ -297,12 +303,17 @@ public class DocUpdateSentinel
                // Inform the user only once if this was an autosave failure.
                if (!loggedAutosaveError_)
                {
-                  loggedAutosaveError_ = true;
+                  // do not show the error if it is a transient autosave related issue - this can occur fairly frequently
+                  // when attempting to save files that are being backed up by external software
+                  if (!message.contains("The process cannot access the file because it is being used by another process"))
+                  {
+                     loggedAutosaveError_ = true;
 
-                  RStudioGinjector.INSTANCE.getGlobalDisplay().showErrorMessage(
-                        "Error Autosaving File",
-                        "RStudio was unable to autosave this file. You may need " +
-                        "to restart RStudio.");
+                     RStudioGinjector.INSTANCE.getGlobalDisplay().showErrorMessage(
+                           "Error Autosaving File",
+                           "RStudio was unable to autosave this file. You may need " +
+                                 "to restart RStudio.");
+                  }
                }
 
                // Use regular completed callback for indicator.
@@ -334,7 +345,7 @@ public class DocUpdateSentinel
 
    public void changeFileType(String fileType, final ProgressIndicator progress)
    {
-      saveWithSuspendedAutoSave(null, fileType, null, progress);
+      saveWithSuspendedAutoSave(null, fileType, null, true, progress);
    }
 
    public void save(String path,
@@ -342,23 +353,25 @@ public class DocUpdateSentinel
                     String fileType,
                     // encoding==null means don't change value
                     String encoding,
+                    boolean retryWrite,
                     final ProgressIndicator progress)
    {
       assert path != null;
       if (path == null)
          throw new IllegalArgumentException("Path cannot be null");
-      saveWithSuspendedAutoSave(path, fileType, encoding, progress);
+      saveWithSuspendedAutoSave(path, fileType, encoding, retryWrite, progress);
    }
 
    private void saveWithSuspendedAutoSave(String path,
                                           String fileType,
                                           String encoding,
+                                          boolean retryWrite,
                                           final ProgressIndicator progress)
    {
       if (autosaver_ != null)
          autosaver_.suspend();
 
-      doSave(path, fileType, encoding, new ProgressIndicator()
+      doSave(path, fileType, encoding, retryWrite, new ProgressIndicator()
       {
          public void onProgress(String message)
          {
@@ -400,12 +413,13 @@ public class DocUpdateSentinel
    private boolean doSave(String path,
                           String fileType,
                           String encoding,
+                          boolean retryWrite,
                           ProgressIndicator progress)
    {
       boolean didSave = false;
       try
       {
-         didSave = doSaveImpl(path, fileType, encoding, progress);
+         didSave = doSaveImpl(path, fileType, encoding, retryWrite, progress);
       }
       catch (Exception ex)
       {
@@ -447,6 +461,7 @@ public class DocUpdateSentinel
    private boolean doSaveImpl(final String path,
                               final String fileType,
                               final String encoding,
+                              final boolean retryWrite,
                               final ProgressIndicator progress)
    {
       /* We need to fork the change tracker so that we can "mark" the moment
@@ -519,6 +534,7 @@ public class DocUpdateSentinel
             diff.getLength(),
             diff.isValid(),
             hash,
+            retryWrite,
             new ServerRequestCallback<String>()
             {
                @Override
@@ -596,7 +612,7 @@ public class DocUpdateSentinel
                   {
                      // We just hit a race condition where two updates
                      // happened at once. Try again
-                     doSave(path, fileType, encoding, progress);
+                     doSave(path, fileType, encoding, retryWrite, progress);
                   }
                   else
                   {
@@ -610,6 +626,7 @@ public class DocUpdateSentinel
                            foldSpec,
                            newChunkDefs,
                            newContents,
+                           retryWrite,
                            this);
                   }
                }